+++ handler/constants.go
@@ -13,29 +13,29 @@ const (
 
 // 角色
 const (
-	GroupOwner int16 = iota + 1 // 群主
-	Admin                       // 管理员
-	Member                      // 普通成员
+	GroupOwner int8 = iota + 1 // 群主
+	Admin                      // 管理员
+	Member                     // 普通成员
 )
 
 // 接收者类型
 const (
-	User  int16 = iota + 1 // 用户
-	Group                  // 群组
+	User  int8 = iota + 1 // 用户
+	Group                 // 群组
 )
 
 // 发送类型
 const (
-	PrivateChat int16 = iota + 1 // 私聊
-	GroupChat                    // 群聊
-	Heartbeat                    // 心跳
+	PrivateChat int8 = iota + 1 // 私聊
+	GroupChat                   // 群聊
+	Heartbeat                   // 心跳
 )
 
 // 消息类型
 const (
-	Text  int16 = iota + 1 // 文本
-	File                   // 文件
-	Image                  // 图片
-	Audio                  // 音频
-	Video                  // 视频
+	Text  int8 = iota + 1 // 文本
+	File                  // 文件
+	Image                 // 图片
+	Audio                 // 音频
+	Video                 // 视频
 )
