+++ handler/validator.go
@@ -8,7 +8,7 @@ import (
 
 // 校验性别是否支持
 var validGender = func(fieldLevel validator.FieldLevel) bool {
-	if gender, ok := fieldLevel.Field().Interface().(int16); ok {
+	if gender, ok := fieldLevel.Field().Interface().(int8); ok {
 		return utils.IsSupportedGender(gender)
 	}
 	return false
